+++ lib/src/query/query_result.dart
@@ -6,7 +6,8 @@ import '../../delta_trace_db.dart';
 /// (ja) DBへのクエリ結果や付加情報を格納したクラスです。
 class QueryResult<T> extends QueryExecutionResult {
   static const String className = "QueryResult";
-  static const String version = "4";
+  static const String version = "5";
+  String target;
   EnumQueryType type;
   List<Map<String, dynamic>> result;
   int dbLength;
@@ -23,15 +24,29 @@ class QueryResult<T> extends QueryExecutionResult {
   /// When false, the value will be true even if the number of updates is 0.
   /// In other cases, if an exception occurs internally,
   /// it will be treated as an error.
+  /// * [target] : The query target collection name in DB.
+  /// This was introduced in version 5 of this class.
+  /// If the server is running an earlier version,
+  /// an empty string is substituted.
   /// * [type] : The query type for this result.
   /// The query type at the time of submission is entered as is.
-  /// * [result] : Search results, update results, deleted objects, etc.
-  /// * [dbLength] : The total number of records in the DB.
+  /// * [result] : The operation result data.
+  ///   Always a list of serialized map objects (JSON).
+  ///   Its meaning depends on [type]:
+  ///   - For "search": The list contains matched objects.
+  ///   - For "add": The list contains the added objects with serial keys.
+  ///   - For "update": The list contains updated objects.
+  ///   - For "delete": The list contains deleted objects.
+  ///   Note: For add/update/delete, the list is only populated if the
+  ///   query option [returnData] is set to true.
+  /// * [dbLength] : DB side item length.
+  /// This is The total number of records in the collection.
   /// * [updateCount] : The total number of records add, updated or deleted.
   /// * [hitCount] : The total number of records searched.
   /// * [errorMessage] : A message that is added only if an error occurs.
   QueryResult({
     required super.isSuccess,
+    required this.target,
     required this.type,
     required this.result,
     required this.dbLength,
@@ -48,6 +63,7 @@ class QueryResult<T> extends QueryExecutionResult {
   factory QueryResult.fromDict(Map<String, dynamic> src) {
     return QueryResult<T>(
       isSuccess: src["isSuccess"],
+      target: src.containsKey("target") ? src["target"] : "",
       type: EnumQueryType.values.byName(src["type"]),
       result: (src["result"] as List).cast<Map<String, dynamic>>(),
       dbLength: src["dbLength"],
@@ -84,6 +100,7 @@ class QueryResult<T> extends QueryExecutionResult {
       "className": className,
       "version": version,
       "isSuccess": isSuccess,
+      "target": target,
       "type": type.name,
       "result": (UtilCopy.jsonableDeepCopy(result) as List)
           .cast<Map<String, dynamic>>(),